Fix @lorenzo45 spaghetti code.

+++ Assignments/Assignment 3/MovieOrganizer/MovieOrganizer/ViewController.m
@@ -205,9 +205,9 @@
 
 #pragma MARK - NSTableView
 
--(int)numberOfRowsInTableView:(NSTableView *)tableView {
+-(NSInteger)numberOfRowsInTableView:(NSTableView *)tableView {
     
-    return (int)self.movies.count;
+    return self.movies.count;
     
 }
